requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.8
django-ipware==6.0.1
idna==3.6
Pillow==10.1.0
python-decouple==3.8
python-ipware==2.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
